+++ packages/yoga/src/TextArea/index.native.js
@@ -1,0 +1,3 @@
+import TextArea from './native';
+
+export default TextArea;