+++ packages/yoga/src/Theme/index.js
@@ -1,4 +1,4 @@
-import ThemeProvider from './Provider';
+import { ThemeProvider, FontLoader } from './Provider';
 import * as themes from './themes';
 
-export { ThemeProvider as default, themes };
+export { ThemeProvider as default, FontLoader, themes };