+++ packages/yoga/src/Theme/themes/components/progress.js
@@ -1,7 +1,6 @@
 const progress = ({ spacing, colors, radii, fontSizes }) => ({
   height: spacing.xsmall,
   backgroundColor: {
-    value: colors.primary[3],
     bar: colors.gray[1],
   },
   border: {