+++ packages/yoga/src/shared/index.js
@@ -1,4 +1,4 @@
 import HiddenInput from './HiddenInput';
-import { limitChildren, typeOf, charLength } from './propTypes';
+import { limitChildren, typeOf, charLength, max } from './propTypes';
 
-export { HiddenInput, limitChildren, typeOf, charLength };
+export { HiddenInput, limitChildren, typeOf, charLength, max };